# [1/N] Update MMQ kernel catch up upstream llama.cpp implementation