--- a/hf-kernels/ggml-kernels/.gitignore
+++ b/hf-kernels/ggml-kernels/.gitignore
@@ -2,6 +2,7 @@
 result
 
 # build2cmake generated files
+cmake/
 CMakeLists.txt
 pyproject.toml
 setup.py