--- a/hf-kernels/ggml-kernels/build.toml
+++ b/hf-kernels/ggml-kernels/build.toml
@@ -14,6 +14,7 @@ src = [
   "ggml/dequantize.cuh",
   "ggml/ggml_kernel.cu",
   "ggml/ggml-common.h",
+  "ggml/mmq.cu",
   "ggml/mmq.cuh",
   "ggml/mmvq.cuh",
   "ggml/vecdotq.cuh",