--- a/hf-kernels/ggml-kernels/ggml/mmq.cu
+++ b/hf-kernels/ggml-kernels/ggml/mmq.cu
@@ -6,6 +6,7 @@
 #include <c10/cuda/CUDAGuard.h>
 
 #include "cuda_compat.h"
+#include "cuda_utils.h"
 #include "dispatch_utils.h"
 
 #include "ggml-common.h"