@Isotr0py 2025-08-09
### Benchmark results
Hardware: NVIDIA GeForce RTX 3090 24GB
Hidden size=1024, num_tokens=4096

| **Quantization** | **Main Time (ms)** | **PR Time (ms)** |
|:----------------:|:------------------:|:----------------:|
|       Q2_K       |     21.37505244    |    9.757954925   |
|       Q3_K       |     54.36249449    |    8.700753478   |
|       Q4_K       |     32.7376706     |    6.185650527   |
|       Q5_K       |     42.49698077    |    6.589026637   |
|       Q6_K       |     45.83637421    |    6.769261481   |
|       Q4_0       |     23.81569172    |    5.485869246   |
|       Q5_0       |     37.47556383    |    5.97816959    |
|       Q8_0       |     40.68143068    |    5.784195224   |